Revert "double instead of float"
This reverts commit 55fdf912fb4c390986b3872f20b2313f96c3b54e.

+++ nine-dof-block/src/main/java/com/robinkirkman/edison/sparkfun/ninedofblock/SFE_LSM9DS0_JNI.java
@@ -140,9 +140,9 @@ public final class SFE_LSM9DS0_JNI {
 		ptr = ctor0(gAddr, xmAddr);
 	}
 	
-	public double gx, gy, gz;
-	public double ax, ay, az;
-	public double mx, my, mz;
+	public float gx, gy, gz;
+	public float ax, ay, az;
+	public float mx, my, mz;
 	public short temperature;
 	
 	private static native long ctor0(int gAddr, int xmAddr);
